+++ CloudClient/inc/Identity.h
@@ -3,7 +3,7 @@
 
 #include "ClientBackend.h"
 
-int Get_Client_Id(Client *c);
-int Set_Client_Id(Client *c, int id);
+unsigned long Get_Client_Id(Client *c);
+int Set_Client_Id(Client *c, unsigned long id);
 
 #endif
